// paperline-backend/types/auth.ts
export interface RegisterPayload {
  firstname: string;
  lastname: string;
  profilePic?: string;
  email: string;
  password: string;
  username: string;
}

export type LoginPayload = Omit<
  RegisterPayload,
  "firstname" | "lastname" | "profilePic" | "username"
>;

export type DecodedUser = {
  id: string;
};